requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.25.0
pyjwt==1.7.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
Flask-Migrate==2.5.3
uwsgi==2.0.19.1
wheel==0.35.1
